Update htmls/favorite.js/button.css
Added a button.css file and linked to html so you can click favorite button when the pokecards populate to toggle the button to favorited and unfavorited. However I havent made a card for the button to be appended to yet when it is rendered on the favorites list page. Also fixed the links so they connect to the root which will help with the smoothness of the website.