Updated all three .htmls to have new tailwind links. Updated favorite.html to have the favorite button on it and removed it from generation.html. Updated the favorite.js to have a card render to the screen with data fetched from the api. Updated the favorite.css to change the rem sizes for the favorites list text.
Updated all three .htmls so they include the correct tailwind links for the Modal if there is more than one. Updated the generation.html to have classes for the buttons so I can know which fetch data I need for each generation and drill through that in javascript however it does not function yet. Also for the favorite.js I added a card that will render to the screen however I cannot seem to get a button to display onto the card for favoriting it, and unfavoriting it, also included the sprite for the pokeman based on which pokemons name is searched up, and have the also added the stats for the pokemon card so when it renders you know what its name is its generation its id and its type.