Updated favorite.js to include working stats list include a function to remove pokemon on toggle button, added an if statement to the toggle button,  and added functionality to the modal that clears the list. Updated favorite.html to include modal for clear list button, and removed the favorites button from the dropdownmenu. Update generations.css to add space. Updated favorite.css to make the button center, and removed some margin from the favorites list header tag. Updated button.css to include indentation.
Updated favorite.js to include working stats list include a function to remove pokemon when toggle button is clicked, added an if statement to the toggle button so that it will loop through the images of starFav and starNoFav to display the correct image when clicked, and added functionality to the modal that clears the list so when you confirm that you want to remove the favorited pokemon from the list it will clear them from the page. Updated favorite.html to include modal for clear list button, and removed the favorites button from the dropdownmenu. Update generations.css to add space. Updated favorite.css to make the button center, and removed some margin from the favorites list header tag so its centered. Updated button.css to include indentation.