Updated favorite.js to include a starFav.png icon that will persist upon clicking it and storing it and then refreshing.
Updated favorite.js to include a starFav.png icon that will persist upon clicking it and storing it and then refreshing. However now when you search pokemon, and they are rendered the code seems to remove the other pokemon that youve searched and rendered and keep only the one youve favorited until you add more but for now you can only add one at a time.