Updated the button.css to remove the style hidden. Updated favorite.css to have a style for the nav buttons, and changed padding size. Updated favorite.js to have comments describing the code, and added a theme changer for the page using flowbite. Updated favorite.html to include a dark/light mode button, and more styles for the page to polish it up. Updated the README.md to include spots for screenshots.
Updated the button.css to remove the style hidden because it was redundant. Updated favorite.css to have a style for the nav buttons so they are seperated, and changed padding size to rem to make it more relevant. Updated favorite.js to have comments describing the code for more clarity, and added a theme changer for the page using flowbite to help polish up ui. Updated favorite.html to include a dark/light mode button for the theme changer, and more styles for the page to polish it up completely. Updated the README.md to include spots for screenshots however I have not added any screenshots.