Updated generation.css to refactor, remove redundancies, and add more media screen resoultions/Updated randomizer.css simple white space below a curly/Updated style.css to include media resolutions for the joke generator, and refactor/remove redundancies/Updated generation.js to prepare for the search bar update and add the dark mode to the page/ Updated script.js to include the dark mode switcher and move a function to be called at the correct place/Updated all the .htmls to change the styling so its polished up.
Updated generation.css to refactor, remove redundancies, and add more media screen resoultions so that it looks polished. Updated randomizer.css simple white space below a curly so that its easier to read. Updated style.css to include media resolutions for the joke generator, and refactor/remove redundancies so that the code is easier to read. Updated generation.js to prepare for the search bar update and add the dark mode to the page so that dark mode can be used on the generations page. Updated script.js to include the dark mode switcher and move a function to be called at the correct place whic helps readability of code. Updated all the .htmls to change the styling so its polished up with theme for each page and a new dark mode color which is similar on each page.