Update changed generation.js to search.js more readible/Updated generation.html to search.html makes it more readible/Updated favorite.js to remove the search button since its a favorite page not a search page/Updated script.js to render pokemon from search and be addable for the favorites page/Updated script.js to remove white space/Updated all the .html files to include an anchor for the new search.html name in the nav bars.
Update changed generation.js to search.js more readible for the user. Updated generation.html to search.html makes it more readible for the user. Updated favorite.js to remove the search button since its a favorite page not a search page that way its entirely polished. Updated script.js to render pokemon from search and be addable for the favorites page so that the favorites page makes sense. Updated script.js to remove white space to make more readible code. Updated all the .html files to include an anchor for the new search.html name in the nav bars so the user can access the page.