Updated generation.css to be called search.css for better readability of code/Updated search.js to include function for the show more button/ Updated search.html to include a button that is wrapped in a div for it to center the button.
Updated generation.css to be called search.css for better readability of code. Updated search.js to include an a function that will be activated by the button and the function lets the pokemon be visible when clicked false so that it can load in the data, then once clicked again it will show a list of pokemon styled similar to the searched pokemon just with less data because its just a list for the show more button that disappears when clicked again. Updated search.html to include a button which is wrapped in a div for it to center.