Updated search.js to remove a redundant toggle pokemonlist function and then remove a redundant code function for the clear list button since its not on the search page and also added a target variable in the async function for the .game_indices so I can display the versionName on the pokecards/Updated search.html so that its got a box for the pokemon list that will be centered/Updated favorite.js to have the same variable and the same targeted data in the string template literal.
Updated search.js to remove a redundant toggle pokemonlist function and then remove a redundant code function for the clear list button since its not on the search page and also added a target variable in the async function for the .game_indices so I can display the versionName on the pokecards so the user knows what pokemon they are looking at. Updated search.html so that its got a box for the pokemon list that will be centered when you click the more pokemon button they were to the left. Updated favorite.js to have the same variable and the same targeted data in the string template literal so the same data capture method using a variable and using a target syntax to grab the version name each time the card was added.